# api/swagger.yml
openapi: 3.0.0
info:
  title: Universal Translator API
  version: 1.0.0
  description: AI-powered translation service for Rocket.Chat
  contact:
    name: API Support
    email: api-support@paysera.tech
servers:
  - url: http://192.168.110.199:3012
    description: Development server
  - url: https://translate-api.paysera.tech
    description: Production server
  - url: https://translate-api-dev.paysera.tech
    description: Development staging server
paths:
  /api/translate:
    post:
      summary: Translate text
      description: Translate text from source language to target language using AI providers
      tags:
        - Translation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
                - targetLang
              properties:
                text:
                  type: string
                  description: Text to translate
                  example: "Hello world"
                  maxLength: 10000
                sourceLang:
                  type: string
                  description: Source language code (auto-detect if not provided)
                  example: "en"
                  minLength: 2
                  maxLength: 10
                targetLang:
                  type: string
                  description: Target language code
                  example: "lt"
                  minLength: 2
                  maxLength: 10
                provider:
                  type: string
                  enum: [openai, anthropic, deepl, google]
                  description: Translation provider
                  default: openai
                context:
                  type: array
                  items:
                    type: string
                  description: Additional context messages for better translation
                  maxItems: 10
                quality:
                  type: string
                  enum: [fast, balanced, quality]
                  description: Translation quality preference
                  default: balanced
                channelId:
                  type: string
                  description: Rocket.Chat channel ID for context
                messageId:
                  type: string
                  description: Rocket.Chat message ID for tracking
      responses:
        200:
          description: Translation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslationResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        429:
          $ref: '#/components/responses/RateLimitExceeded'
        500:
          $ref: '#/components/responses/InternalServerError'
...
  /api/detect-language:
    post:
      summary: Detect language
      description: Automatically detect the language of provided text
      tags:
        - Language Detection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Text to analyze
                  example: "Bonjour le monde"
                  maxLength: 5000
      responses:
        200:
          description: Language detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LanguageDetectionResponse'
        400:
          $ref: '#/components/responses/BadRequest'
...
components:
  schemas:
    TranslationResponse:
      type: object
      properties:
        translatedText:
          type: string
          description: Translated text result
          example: "Labas pasauli"
        sourceLang:
          type: string
          description: Detected or provided source language
          example: "en"
        targetLang:
          type: string
          description: Target language
          example: "lt"
        confidence:
          type: number
          format: float
          description: Translation confidence score (0-1)
          example: 0.95
          minimum: 0
          maximum: 1
        provider:
          type: string
          description: Provider used for translation
          example: "openai"
        model:
          type: string
          description: Specific model used by provider
          example: "gpt-4"
        cached:
          type: boolean
          description: Whether result was from cache
          example: false
        cost:
          type: number
          format: float
          description: Cost in USD for this translation
          example: 0.002
        processingTime:
          type: integer
          description: Processing time in milliseconds
          example: 1250
...
    LanguageDetectionResponse:
      type: object
      properties:
        language:
          type: string
          description: Detected language code
          example: "fr"
        confidence:
          type: number
          format: float
          description: Detection confidence (0-1)
          example: 0.98
          minimum: 0
          maximum: 1
        alternatives:
          type: array
          items:
            type: object
            properties:
              language:
                type: string
                description: Alternative language code
              confidence:
                type: number
                format: float
                description: Confidence for this alternative
          description: Alternative language detections with lower confidence
...
  responses:
    BadRequest:
      description: Bad request - Invalid parameters
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid language code"
              code:
                type: string
                example: "INVALID_LANGUAGE"
              details:
                type: object
                description: Additional error details
    Unauthorized:
      description: Unauthorized - Invalid or missing API key
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid API key"
              code:
                type: string
                example: "UNAUTHORIZED"
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Rate limit exceeded"
              retryAfter:
                type: integer
                description: Seconds until next request allowed
                example: 60
              limit:
                type: integer
                description: Requests per time window
              remaining:
                type: integer
                description: Remaining requests in current window
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error"
              requestId:
                type: string
                description: Request ID for debugging
                example: "req_1234567890"
              code:
                type: string
                example: "INTERNAL_ERROR"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    InternalAuth:
      type: apiKey
      in: header
      name: X-Internal-Secret
      description: Internal service authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - InternalAuth: []
...